api_yamdb/requirements.txt
asgiref==3.5.2
Django==3.2
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.6
sqlparse==0.4.3
gunicorn==20.0.4
psycopg2-binary==2.8.6
